Move projects into separate folder
