+++ src/App.js
@@ -1,7 +1,7 @@
 import About from "./pages/About/About";
 import Contact from "./pages/Contact";
 import Home from "./pages/Home/Home";
-import Projects from "./pages/Projects";
+import Projects from "./pages/Projects/Projects";
 
 import "./App.css";
